Dashboard header adjustments
